Move interface to its own file, keep lib.rs clean
